push1

+++ src/main/java/com/cognizant/inventory_service/repository/AllotmentRepository.java
@@ -1,9 +1,11 @@
 package com.cognizant.inventory_service.repository;
  
-import com.cognizant.inventory_service.entity.Allotment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.cognizant.inventory_service.entity.Allotment;
  
 @Repository
-public interface AllotmentRepository extends JpaRepository<Allotment, Long> {
+public interface AllotmentRepository extends JpaRepository<Allotment, Integer> {
+	
 }
